feat(api): support carryforward flags on parallel finish
Split the oversized cli_test.py into a per-subcommand tests/cli/ package
and the CI-detection cases out of resolve_test.py so both stay under the
max-module-lines limit as they grow.

changelog:
``coveralls finish`` now accepts ``--carryforward`` (also
``COVERALLS_CARRYFORWARD`` or the ``carryforward`` config-file key) to
carry forward coverage flags for jobs missing from a parallel build.
